Refactoring out Datasets into own scripts so that they can be used without pytorch lightning adding default tokenizer to utils as it is used in all VQA DataSets
