+++ configilm/extra/DataSets/RSVQAxBEN_DataSet.py
@@ -9,28 +9,17 @@ https://bigearth.net/
 """
 import json
 import os
-from datetime import datetime
 from os.path import join
-from pathlib import Path
-from time import time
-from typing import List
 from typing import Optional
-from typing import Union
 
 import numpy as np
 import psutil
-import pytorch_lightning as pl
 import torch
-import torchvision.transforms as transforms
-from torch.utils.data import DataLoader
 from torch.utils.data import Dataset
 from tqdm import tqdm
-from transformers import BertTokenizer
 
-from configilm.extra.BEN_lmdb_utils import band_combi_to_mean_std
 from configilm.extra.BEN_lmdb_utils import BENLMDBReader
-from configilm.extra.CustomTorchClasses import MyGaussianNoise
-from configilm.extra.CustomTorchClasses import MyRotateTransform
+from configilm.util import get_default_tokenizer
 from configilm.util import huggingface_tokenize_and_pad
 from configilm.util import Messages
 
@@ -114,16 +103,7 @@ class RSVQAxBENDataSet(Dataset):
                 "result in very bad performance if the used network expected other "
                 "tokens."
             )
-            # get path relative to this script, not relative to the calling main script
-            # TODO: make this more flexible
-            default_tokenizer = (
-                Path(__file__)
-                .parent.parent.joinpath(
-                    "huggingface_tokenizers", "bert-base-uncased.tok"
-                )
-                .resolve(True)
-            )
-            self.tokenizer = BertTokenizer.from_pretrained(default_tokenizer.resolve())
+            self.tokenizer = get_default_tokenizer()
         else:
             self.tokenizer = tokenizer
         self.seq_length = seq_length
@@ -241,167 +221,3 @@ class RSVQAxBENDataSet(Dataset):
         )
 
         return img, question_ids, label
-
-
-class RSVQAxBENDataModule(pl.LightningDataModule):
-    train_ds: Union[None, RSVQAxBENDataSet] = None
-    val_ds: Union[None, RSVQAxBENDataSet] = None
-    test_ds: Union[None, RSVQAxBENDataSet] = None
-    selected_answers: Union[None, List[str]] = None
-
-    def __init__(
-        self,
-        batch_size=16,
-        data_dir: str = "./",
-        img_size=None,
-        num_workers_dataloader=None,
-        max_img_idx=None,
-        shuffle=None,
-        tokenizer=None,
-        seq_length=32,
-        selected_answers=None,
-        pin_memory=None,
-    ):
-        if img_size is not None and len(img_size) != 3:
-            raise ValueError(
-                f"Expected image_size with 3 dimensions (HxWxC) or None but got "
-                f"{len(img_size)} dimensions instead"
-            )
-        super().__init__()
-        if num_workers_dataloader is None:
-            cpu_count = os.cpu_count()
-            if type(cpu_count) is int:
-                self.num_workers_dataloader = cpu_count // 2
-            else:
-                self.num_workers_dataloader = 0
-        else:
-            self.num_workers_dataloader = num_workers_dataloader
-        print(f"Dataloader using {self.num_workers_dataloader} workers")
-
-        self.data_dir = data_dir
-        self.batch_size = batch_size
-        self.max_img_idx = (
-            max_img_idx if max_img_idx is None or max_img_idx > 0 else None
-        )
-        self.img_size = (12, 120, 120) if img_size is None else img_size
-        self.shuffle = shuffle
-        if self.shuffle is not None:
-            Messages.hint(
-                f"Shuffle was set to {self.shuffle}. This is not recommended for most "
-                f"configuration. Use shuffle=None (default) for recommended "
-                f"configuration."
-            )
-        self.selected_answers = selected_answers
-
-        ben_mean, ben_std = band_combi_to_mean_std(self.img_size[0])
-
-        self.train_transform = transforms.Compose(
-            [
-                transforms.Resize((self.img_size[1], self.img_size[2]), antialias=True),
-                MyGaussianNoise(20),
-                transforms.RandomHorizontalFlip(),
-                transforms.RandomVerticalFlip(),
-                MyRotateTransform([0, 90, 180, 270]),
-                transforms.Normalize(ben_mean, ben_std),
-            ]
-        )
-        self.transform = transforms.Compose(
-            [
-                transforms.Resize((self.img_size[1], self.img_size[2]), antialias=True),
-                transforms.Normalize(ben_mean, ben_std),
-            ]
-        )
-        # self.transform = None
-        self.pin_memory = torch.cuda.device_count() > 0
-        self.pin_memory = self.pin_memory if pin_memory is None else pin_memory
-        Messages.hint(
-            f"pin_memory set to {pin_memory}"
-            f"{' ' if pin_memory is None else ' via overwrite'}"
-        )
-
-        self.tokenizer = tokenizer
-        self.seq_length = seq_length
-
-    def prepare_data(self):
-        pass
-
-    def setup(self, stage: Optional[str] = None):
-        print(f"({datetime.now().strftime('%H:%M:%S')}) Datamodule setup called")
-        sample_info_msg = ""
-        t0 = time()
-
-        # Assign train/val datasets for use in dataloaders
-        if stage == "fit" or stage is None:
-            if self.train_ds is None:
-                self.train_ds = RSVQAxBENDataSet(
-                    self.data_dir,
-                    split="train",
-                    transform=self.train_transform,
-                    max_img_idx=self.max_img_idx,
-                    img_size=self.img_size,
-                    tokenizer=self.tokenizer,
-                    seq_length=self.seq_length,
-                )
-            if self.selected_answers is None:
-                self.selected_answers = self.train_ds.selected_answers
-
-            if self.val_ds is None:
-                self.val_ds = RSVQAxBENDataSet(
-                    self.data_dir,
-                    split="val",
-                    transform=self.transform,
-                    max_img_idx=self.max_img_idx,
-                    img_size=self.img_size,
-                    tokenizer=self.tokenizer,
-                    seq_length=self.seq_length,
-                    selected_answers=self.selected_answers,
-                )
-            sample_info_msg += f"  Total training samples: {len(self.train_ds):8,d}"
-            sample_info_msg += f"  Total validation samples: {len(self.val_ds):8,d}"
-
-        # Assign test dataset for use in dataloader(s)
-        if (stage == "test" or stage is None) and self.test_ds is None:
-            self.test_ds = RSVQAxBENDataSet(
-                self.data_dir,
-                split="test",
-                transform=self.transform,
-                max_img_idx=self.max_img_idx,
-                img_size=self.img_size,
-                tokenizer=self.tokenizer,
-                seq_length=self.seq_length,
-                selected_answers=self.selected_answers,
-            )
-            sample_info_msg += f"  Total test samples: {len(self.test_ds):8,d}"
-
-        if stage == "predict":
-            raise NotImplementedError("Predict stage not implemented")
-
-        print(f"setup took {time() - t0:.2f} seconds")
-        print(sample_info_msg)
-
-    def train_dataloader(self):
-        return DataLoader(
-            self.train_ds,
-            batch_size=self.batch_size,
-            shuffle=True if self.shuffle is None else self.shuffle,
-            num_workers=self.num_workers_dataloader,
-            pin_memory=self.pin_memory,
-        )
-
-    def val_dataloader(self):
-        return DataLoader(
-            self.val_ds,
-            batch_size=self.batch_size,
-            shuffle=False if self.shuffle is None else self.shuffle,
-            num_workers=self.num_workers_dataloader,
-            pin_memory=self.pin_memory,
-        )
-
-    def test_dataloader(self):
-        return DataLoader(
-            self.test_ds,
-            batch_size=self.batch_size,
-            shuffle=False if self.shuffle is None else self.shuffle,
-            num_workers=self.num_workers_dataloader,
-            pin_memory=self.pin_memory,
-        )
