+++ configilm/extra/huggingface_tokenizers/bert-base-uncased.tok/vocab.txt
